Implements a role for installing Xcode and deps.
This uses parts of: https://github.com/edunham/ansible-role-xcode

But fixes an issue with installing recent versions of Xcode (a hard link issue)
and also installs the full set of dependencies for Xcode (avoiding an alert on
initial run).

The Xcode disk image should be fetched separately (it's held behind a login
screen) and so this is ignored.